// src/app/models/Historical.ts
export class Historical {
  id?: number;
  courseId: number;
  userId: number;
  time: string;
  learned?: number;
  learnTime?: string;
  learnProgress?: string;
  notesNum?: number;
  questionNum?: number;
  statue?: number;
  learnedSection?: string;
  learningHalf?: string;
  newLearn?: number;
}